Doctor.money Doctor.secondMoney 从 int 类型 改为 double 类型

+++ app/src/main/java/com/doctor/sun/entity/AppointmentBuilder.java
@@ -199,7 +199,7 @@ public class AppointmentBuilder extends BaseObservable implements Parcelable {
         context.startActivity(intent);
     }
 
-    public int money() {
+    public double money() {
         switch (getType()) {
             case AppointmentType.STANDARD:
                 return doctor.getSecondMoney();
